+++ opensand-core/src/conf/OpenSandModelConf.h
@@ -172,12 +172,12 @@ public:
 	bool getLocalStorage(bool &enabled, std::string &output_folder) const;
 	bool getRemoteStorage(bool &enabled,
 	                      std::string &address,
-	                      unsigned short &stats_port,
-	                      unsigned short &logs_port) const;
+	                      uint16_t &stats_port,
+	                      uint16_t &logs_port) const;
 	bool logLevels(std::map<std::string, log_level_t> &levels) const;
 	bool getSarp(SarpTable &sarp_table) const;
-	bool getNccPorts(int &pep_tcp_port, int &svno_tcp_port) const;
-	bool getQosServerHost(std::string &qos_server_host_agent, int &qos_server_host_port) const;
+	bool getNccPorts(uint16_t &pep_tcp_port, uint16_t &svno_tcp_port) const;
+	bool getQosServerHost(std::string &qos_server_host_agent, uint16_t &qos_server_host_port) const;
 	bool getS2WaveFormsDefinition(std::vector<fmt_definition_parameters> &fmt_definitions) const;
 	bool getRcs2WaveFormsDefinition(std::vector<fmt_definition_parameters> &fmt_definitions,
 	                                vol_sym_t req_burst_length) const;
@@ -206,8 +206,8 @@ public:
 	bool getSpotForwardCarriers(tal_id_t gw_id, OpenSandModelConf::spot &spot) const;
 	bool getInterconnectCarrier(bool upward_connection,
 	                            std::string &remote_address,
-	                            unsigned int &data_port,
-	                            unsigned int &sig_port,
+	                            uint16_t &data_port,
+	                            uint16_t &sig_port,
 	                            unsigned int &udp_stack,
 	                            unsigned int &udp_rmem,
 	                            unsigned int &udp_wmem,
